Add strict `NonEmptyMap` type and tests.
This change introduces a `NonEmptyMap` type with strict semantics.

The implementation is based on the implementation of 'Data.Map.Strict'
provided by the 'containers' package, but provides an extra guarantee
that the map contains at least one entry at all times.

At the time of writing, there are other similar types available on GitHub, but
these implementations all extend the lazy `Data.Map` type, and thus provide
lazy semantics.

However, in the case of long-lived data structures, a strict type is more
useful, as it allows us to avoid building up large numbers of unevaluated
thunks.