Prevent export of default data constructor for `TokenMap`.
This change adds a test to check that the default data constructor for
`TokenMap` is not exported.

The default data constructor for 'TokenMap' is not exported, by design,
as the internal data structure has an invariant that must be preserved
across all operations.

Exporting the default constructor would make it possible for functions
outside the 'TokenMap' module to break the invariant, opening the door
to subtle regressions.