Prevent export of default data constructor for `NonEmptyMap`.
This change adds a test to check that the default data constructor for
`NonEmptyMap` is not exported.

The default data constructor for 'NonEmptyMap' is not exported, by
design, as the internal data structure has an invariant that must be
preserved across all operations.

Exporting the default constructor would make it possible for functions
in other modules to break the invariant, opening the door to subtle
regressions.